requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Session==0.5.0

# Security
bcrypt==4.1.2
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

# Forms and validation
WTForms==3.1.1
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.2
pytest-flask==1.3.0

# Development
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Web Push Notifications
pywebpush==2.3.0
APScheduler==3.11.2
